Refactored stream service
