+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/stream/InMemoryStreamCoordinatorClient.java
@@ -19,7 +19,6 @@ import co.cask.cdap.api.data.stream.StreamSpecification;
 import co.cask.cdap.common.conf.InMemoryPropertyStore;
 import co.cask.cdap.common.conf.PropertyStore;
 import co.cask.cdap.common.io.Codec;
-import co.cask.cdap.data.stream.service.StreamCoordinator;
 import co.cask.cdap.data.stream.service.StreamMetaStore;
 import co.cask.cdap.data2.transaction.stream.StreamAdmin;
 import com.google.common.base.Throwables;
@@ -29,8 +28,6 @@ import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import org.apache.twill.common.Cancellable;
-import org.apache.twill.discovery.Discoverable;
 
 import java.util.Set;
 import javax.annotation.Nullable;
@@ -39,9 +36,7 @@ import javax.annotation.Nullable;
  * In memory implementation for {@link StreamCoordinatorClient}.
  */
 @Singleton
-public final class InMemoryStreamCoordinatorClient
-  extends AbstractStreamCoordinatorClient
-  implements StreamCoordinator {
+public final class InMemoryStreamCoordinatorClient extends AbstractStreamCoordinatorClient {
 
   private final StreamMetaStore streamMetaStore;
   private final Set<StreamLeaderListener> leaderListeners;
@@ -66,35 +61,6 @@ public final class InMemoryStreamCoordinatorClient
   @Override
   protected <T> PropertyStore<T> createPropertyStore(Codec<T> codec) {
     return new InMemoryPropertyStore<T>();
-  }
-
-  @Override
-  public void setHandlerDiscoverable(Discoverable discoverable) {
-    // No-op
-  }
-
-  @Override
-  public Cancellable addLeaderListener(final StreamLeaderListener listener) {
-    // Create a wrapper around user's listener, to ensure that the cancelling behavior set in this method
-    // is not overridden by user's code implementation of the equal method
-    final StreamLeaderListener wrappedListener = new StreamLeaderListener() {
-      @Override
-      public void leaderOf(Set<String> streamNames) {
-        listener.leaderOf(streamNames);
-      }
-    };
-
-    synchronized (this) {
-      leaderListeners.add(wrappedListener);
-    }
-    return new Cancellable() {
-      @Override
-      public void cancel() {
-        synchronized (InMemoryStreamCoordinatorClient.this) {
-          leaderListeners.remove(wrappedListener);
-        }
-      }
-    };
   }
 
   @Override
