+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/stream/service/heartbeat/StreamWriterHeartbeat.java
@@ -16,12 +16,12 @@
 
 package co.cask.cdap.data.stream.service.heartbeat;
 
-import co.cask.cdap.data.stream.service.StreamCoordinator;
+import co.cask.cdap.data.stream.service.DistributedStreamService;
 import com.google.common.base.Objects;
 
 /**
  * Heartbeat sent by a Stream writer containing the total size of its files, in bytes. The heartbeats
- * concerning one stream are aggregated by the {@link StreamCoordinator} elected leader of the stream.
+ * concerning one stream are aggregated by the {@link DistributedStreamService} elected leader of the stream.
  */
 public class StreamWriterHeartbeat {
