+++ cdap-standalone/src/main/java/co/cask/cdap/StandaloneMain.java
@@ -32,7 +32,7 @@ import co.cask.cdap.data.runtime.DataFabricModules;
 import co.cask.cdap.data.runtime.DataSetServiceModules;
 import co.cask.cdap.data.runtime.DataSetsModules;
 import co.cask.cdap.data.stream.StreamAdminModules;
-import co.cask.cdap.data.stream.service.StreamHttpService;
+import co.cask.cdap.data.stream.service.LocalStreamService;
 import co.cask.cdap.data.stream.service.StreamServiceRuntimeModule;
 import co.cask.cdap.data2.datafabric.dataset.service.DatasetService;
 import co.cask.cdap.explore.client.ExploreClient;
@@ -81,9 +81,8 @@ public class StandaloneMain {
   private final NettyRouter router;
   private final MetricsQueryService metricsQueryService;
   private final AppFabricServer appFabricServer;
-  private final StreamHttpService streamHttpService;
   private final ServiceStore serviceStore;
-
+  private final LocalStreamService streamService;
   private final MetricsCollectionService metricsCollectionService;
 
   private final LogAppenderInitializer logAppenderInitializer;
@@ -111,10 +110,9 @@ public class StandaloneMain {
     metricsCollectionService = injector.getInstance(MetricsCollectionService.class);
     datasetService = injector.getInstance(DatasetService.class);
     serviceStore = injector.getInstance(ServiceStore.class);
+    streamService = injector.getInstance(LocalStreamService.class);
 
     this.webCloudAppService = (webAppPath == null) ? null : injector.getInstance(WebCloudAppService.class);
-
-    streamHttpService = injector.getInstance(StreamHttpService.class);
 
     sslEnabled = configuration.getBoolean(Constants.Security.SSL_ENABLED);
     securityEnabled = configuration.getBoolean(Constants.Security.ENABLED);
@@ -154,6 +152,7 @@ public class StandaloneMain {
     metricsCollectionService.startAndWait();
     datasetService.startAndWait();
     serviceStore.startAndWait();
+    streamService.startAndWait();
 
     // It is recommended to initialize log appender after datasetService is started,
     // since log appender instantiates a dataset.
@@ -169,7 +168,6 @@ public class StandaloneMain {
     if (webCloudAppService != null) {
       webCloudAppService.startAndWait();
     }
-    streamHttpService.startAndWait();
 
     if (securityEnabled) {
       externalAuthenticationServer.startAndWait();
@@ -202,7 +200,7 @@ public class StandaloneMain {
       //  shut down router to stop all incoming traffic
       router.stopAndWait();
       // now the stream writer and the explore service (they need tx)
-      streamHttpService.stopAndWait();
+      streamService.stopAndWait();
       if (exploreExecutorService != null) {
         exploreExecutorService.stopAndWait();
       }
